+++ CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [1.16.0] - 2018-08-24
+### Changed
+- `SearchBar` design.
+- Update `Header` design to a white style.
+
 ### Fix
 - Component name into the documentation of `CategoryHighlights`.